# Improve performance of decrypt one shots