--- a/src/libraries/Common/tests/System/Security/Cryptography/AlgorithmImplementations/Symmetric/SymmetricOneShotBase.cs
+++ b/src/libraries/Common/tests/System/Security/Cryptography/AlgorithmImplementations/Symmetric/SymmetricOneShotBase.cs
@@ -597,5 +597,12 @@ namespace System.Security.Cryptography.Tests
                 AssertExtensions.SequenceEqual(expected, actual);
             }
         }
+
+        protected static byte[] ByteArrayFilledWith(byte contents, int length)
+        {
+            byte[] ret = new byte[length];
+            ret.AsSpan().Fill(contents);
+            return ret;
+        }
     }
 }