* If the destination is big enough, even for padded modes, decrypt directly to the user supplied buffer. If the padding is invalid, then the data written to the buffer is zeroed. If the padding is valid, then the padding bytes are zeroed so the padding is not "beyond" what was reported in `bytesWritten`.
* If the destination is impossibly small, i.e. the buffer is too short by more than one block, bail out early.
* If the plaintext is sufficiently small, decrypt in to a stack buffer and copy to the destination if the padding is valid.
* Otherwise, decrypt all but the last block directly to the user buffer. The last block is decrypted in to a stack buffer. If the last block does not fit in the user buffer, zero everything that was written so far. Otherwise, copy the last block to the user buffer.

Contributes to #55601

Remaining work:

- [x] Better unit tests on boundary conditions
- [x] Gather performance metrics for all platforms under all conditions