@vcsjones 2022-02-22
Ohhh. We need to handle overlapping source / destination. I... think... if the source and destination overlap we go back to the original behavior and do everything in a rented buffer.