@bartonjs 2022-02-26
> @bartonjs early feedback welcome

Early feedback is ✔️- nice work.  (Few nits/questions)