@vcsjones 2022-03-10
Benchmarks show a decent improvement that scales with the size of the ciphertext to be decrypted for padded data.

<details>

`GuaranteedBufferSize` tests are where the `destination` is `>=` to the input size, even with padding, so the destination is unambiguously big enough.

`ExactBufferSize` is where the data is padded, and the destination is big enough to hold the de-padded data, but we need to two-step the decrypt because we don't know how big the padding is.

<pre>
|                           Method |        Job |      Toolchain | DataSize |  Padding |                                     Algorithm |             Mean |           Error |          StdDev |           Median | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 |   Allocated |
|--------------------------------- |----------- |--------------- |--------- |--------- |---------------------------------------------- |-----------------:|----------------:|----------------:|-----------------:|------:|--------:|-------:|------:|------:|------------:|
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |       16 |    PKCS7 |       Internal.Cryptography.AesImplementation |         586.7 ns |         4.01 ns |         3.75 ns |         585.6 ns |  0.99 |    0.01 | 0.0200 |     - |     - |       128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |       16 |    PKCS7 |       Internal.Cryptography.AesImplementation |         593.2 ns |         5.78 ns |         5.41 ns |         591.3 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |       16 |    PKCS7 |       Internal.Cryptography.AesImplementation |         593.6 ns |         5.07 ns |         4.74 ns |         595.4 ns |  0.98 |    0.01 | 0.0200 |     - |     - |       128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |       16 |    PKCS7 |       Internal.Cryptography.AesImplementation |         606.1 ns |         6.77 ns |         6.34 ns |         607.9 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |       16 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      10,873.0 ns |        29.89 ns |        26.49 ns |      10,864.8 ns |  0.98 |    0.01 | 0.0153 |     - |     - |       168 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |       16 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      11,128.9 ns |       156.99 ns |       131.09 ns |      11,182.6 ns |  1.00 |    0.00 |      - |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |       16 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      11,041.7 ns |       214.22 ns |       263.08 ns |      10,922.9 ns |  1.03 |    0.03 | 0.0153 |     - |     - |       168 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |       16 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      10,794.1 ns |        72.09 ns |        63.91 ns |      10,780.2 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |       16 | ANSIX923 |       Internal.Cryptography.AesImplementation |         588.3 ns |         4.61 ns |         4.32 ns |         589.5 ns |  0.99 |    0.01 | 0.0200 |     - |     - |       128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |       16 | ANSIX923 |       Internal.Cryptography.AesImplementation |         591.1 ns |         2.46 ns |         1.92 ns |         590.8 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |       16 | ANSIX923 |       Internal.Cryptography.AesImplementation |         586.1 ns |         9.79 ns |         9.16 ns |         581.0 ns |  0.97 |    0.02 | 0.0200 |     - |     - |       128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |       16 | ANSIX923 |       Internal.Cryptography.AesImplementation |         603.1 ns |         4.13 ns |         3.86 ns |         605.4 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |       16 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      10,692.8 ns |        43.37 ns |        38.45 ns |      10,696.6 ns |  0.98 |    0.01 | 0.0153 |     - |     - |       168 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |       16 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      10,910.8 ns |        41.72 ns |        39.03 ns |      10,901.8 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |       16 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      10,672.5 ns |        44.21 ns |        41.35 ns |      10,656.4 ns |  0.98 |    0.00 | 0.0153 |     - |     - |       168 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |       16 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      10,872.0 ns |        53.10 ns |        49.67 ns |      10,880.8 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      112 |    PKCS7 |       Internal.Cryptography.AesImplementation |         593.4 ns |         8.19 ns |         7.66 ns |         597.0 ns |  1.00 |    0.02 | 0.0200 |     - |     - |       128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      112 |    PKCS7 |       Internal.Cryptography.AesImplementation |         593.4 ns |         8.45 ns |         7.90 ns |         591.4 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      112 |    PKCS7 |       Internal.Cryptography.AesImplementation |         579.3 ns |         7.56 ns |         7.07 ns |         575.9 ns |  0.97 |    0.02 | 0.0200 |     - |     - |       128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |      112 |    PKCS7 |       Internal.Cryptography.AesImplementation |         598.7 ns |         8.63 ns |         8.07 ns |         600.9 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      112 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      13,727.3 ns |        54.78 ns |        51.24 ns |      13,707.6 ns |  1.01 |    0.01 | 0.0153 |     - |     - |       168 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      112 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      13,576.4 ns |        34.84 ns |        29.09 ns |      13,577.5 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      112 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      13,499.5 ns |        56.73 ns |        53.07 ns |      13,505.4 ns |  1.01 |    0.01 | 0.0153 |     - |     - |       168 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |      112 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      13,409.7 ns |        48.25 ns |        40.29 ns |      13,395.6 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      112 | ANSIX923 |       Internal.Cryptography.AesImplementation |         592.3 ns |         9.91 ns |         9.27 ns |         597.3 ns |  0.99 |    0.02 | 0.0200 |     - |     - |       128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      112 | ANSIX923 |       Internal.Cryptography.AesImplementation |         598.6 ns |         7.32 ns |         6.84 ns |         602.1 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      112 | ANSIX923 |       Internal.Cryptography.AesImplementation |         582.7 ns |        10.65 ns |         9.97 ns |         579.4 ns |  0.97 |    0.03 | 0.0200 |     - |     - |       128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |      112 | ANSIX923 |       Internal.Cryptography.AesImplementation |         598.0 ns |         7.18 ns |         6.72 ns |         601.7 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      112 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      13,477.2 ns |        39.64 ns |        37.08 ns |      13,468.7 ns |  0.99 |    0.01 | 0.0153 |     - |     - |       168 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      112 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      13,555.5 ns |        64.38 ns |        53.76 ns |      13,545.9 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      112 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      13,678.1 ns |        41.93 ns |        39.22 ns |      13,691.4 ns |  0.99 |    0.01 | 0.0153 |     - |     - |       168 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |      112 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      13,835.9 ns |       183.36 ns |       171.52 ns |      13,828.6 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      113 |    PKCS7 |       Internal.Cryptography.AesImplementation |         592.5 ns |         1.80 ns |         1.68 ns |         592.1 ns |  1.01 |    0.02 | 0.0200 |     - |     - |       128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      113 |    PKCS7 |       Internal.Cryptography.AesImplementation |         587.3 ns |        10.78 ns |        10.09 ns |         581.3 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      113 |    PKCS7 |       Internal.Cryptography.AesImplementation |         586.1 ns |         6.34 ns |         5.93 ns |         586.9 ns |  0.98 |    0.03 | 0.0200 |     - |     - |       128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |      113 |    PKCS7 |       Internal.Cryptography.AesImplementation |         597.5 ns |        11.53 ns |        15.39 ns |         593.1 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      113 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      13,729.4 ns |       273.20 ns |       408.92 ns |      13,577.3 ns |  0.98 |    0.04 | 0.0153 |     - |     - |       168 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      113 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      14,002.6 ns |       271.15 ns |       361.97 ns |      13,830.1 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      113 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      13,715.9 ns |       265.98 ns |       381.46 ns |      13,602.4 ns |  1.00 |    0.03 | 0.0153 |     - |     - |       168 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |      113 |    PKCS7 | Internal.Cryptography.TripleDesImplementation |      13,818.6 ns |       149.67 ns |       132.68 ns |      13,787.0 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      113 | ANSIX923 |       Internal.Cryptography.AesImplementation |         590.5 ns |         3.26 ns |         2.72 ns |         590.5 ns |  0.98 |    0.01 | 0.0200 |     - |     - |       128 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      113 | ANSIX923 |       Internal.Cryptography.AesImplementation |         601.5 ns |         4.48 ns |         4.19 ns |         604.0 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      113 | ANSIX923 |       Internal.Cryptography.AesImplementation |         597.1 ns |         3.82 ns |         3.57 ns |         598.7 ns |  0.97 |    0.01 | 0.0200 |     - |     - |       128 B |
|      Decrypt_Cbc_ExactBufferSize | Job-ZDOVDV | main @ febaf76 |      113 | ANSIX923 |       Internal.Cryptography.AesImplementation |         612.5 ns |         7.67 ns |         7.18 ns |         610.7 ns |  1.00 |    0.00 | 0.0200 |     - |     - |       128 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
| Decrypt_Cbc_GuaranteedBufferSize | Job-JCRJRQ |         branch |      113 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      13,524.4 ns |        43.52 ns |        40.71 ns |      13,510.1 ns |  0.98 |    0.01 | 0.0153 |     - |     - |       168 B |
| Decrypt_Cbc_GuaranteedBufferSize | Job-ZDOVDV | main @ febaf76 |      113 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      13,811.2 ns |       151.69 ns |       118.43 ns |      13,798.0 ns |  1.00 |    0.00 | 0.0153 |     - |     - |       168 B |
|                                  |            |                |          |          |                                               |                  |                 |                 |                  |       |         |        |       |       |             |
|      Decrypt_Cbc_ExactBufferSize | Job-JCRJRQ |         branch |      113 | ANSIX923 | Internal.Cryptography.TripleDesImplementation |      13,484.9 ns |        16.70 ns |        13.04 ns |      13,485.8 ns |  0.99 |    0.00 | 0.0153 |     - |     - |       168 B |
…